requirements.txt
streamlit
google-generativeai
python-dotenv
pdf2image
matplotlib
seaborn
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai